// src/enums/Theme/paletteEnum.ts
export enum Palette {
  White = "#ffffff",
  Black = "#000000",
  Divider = "rgba(17, 17, 17, 0.15)",
  Gray50 = "#fafafa",
  Gray100 = "#f4f4f5",
  Gray200 = "#e4e4e7",
  Gray300 = "#d4d4d8",
  Gray400 = "#a1a1aa",
  Gray500 = "#71717a",
  Gray600 = "#52525b",
  Gray700 = "#3f3f46",
  Gray800 = "#27272a",
  Gray900 = "#18181b",
  Blue50 = "#e8f8ff",
  Blue100 = "#bee3f8",
  Blue200 = "#90cdf4",
  Blue300 = "#63b3ed",
  Blue400 = "#4299e1",
  Blue500 = "#3182ce",
  Blue600 = "#2b6cb0",
  Blue700 = "#2c5282",
  Blue800 = "#2a4365",
  Blue900 = "#1a365d",
  Green50 = "#f0fff4",
  Green100 = "#c6f6d5",
  Green200 = "#9ae6b4",
  Green300 = "#68d391",
  Green400 = "#48bb78",
  Green500 = "#38a169",
  Green600 = "#25855a",
  Green700 = "#276749",
  Green800 = "#22543d",
  Green900 = "#1c4532",
  Yellow50 = "#fefce8",
  Yellow100 = "#fdedd3",
  Yellow200 = "#fbdba7",
  Yellow300 = "#f9c97c",
  Yellow400 = "#f7b750",
  Yellow500 = "#f5a524",
  Yellow600 = "#c4841d",
  Yellow700 = "#936316",
  Yellow800 = "#62420e",
  Yellow900 = "#312107",
  Red50 = "#ffff5f",
  Red100 = "#fed7d7",
  Red200 = "#feb2b2",
  Red300 = "#fc8181",
  Red400 = "#f56565",
  Red500 = "#e53e3e",
  Red600 = "#c53030",
  Red700 = "#9b2c2c",
  Red800 = "#822727",
  Red900 = "#63171b",
  Purple50 = "#faf5ff",
  Purple100 = "#e9d8fd",
  Purple200 = "#d6bcfa",
  Purple300 = "#b794f4",
  Purple400 = "#9f7aea",
  Purple500 = "#805ad5",
  Purple600 = "#6b46c1",
  Purple700 = "#553c9a",
  Purple800 = "#44337a",
  Purple900 = "#322659",
}